+++ Tests/VeinTests/Encryption/EncryptionTest.swift
@@ -13,9 +13,11 @@
 import Foundation
 import Testing
 #if TEST_SWIFTUI
-    @_spi(VeinTesting) @testable import VeinSwiftUI
-#elseif !TEST_SWIFTUI
-    @_spi(VeinTesting) @testable import VeinCore
+@_spi(VeinTesting) @testable import VeinSwiftUI
+#elseif TEST_SCUI
+@_spi(VeinTesting) @testable import VeinSCUI
+#else
+@_spi(VeinTesting) @testable import VeinCore
 #endif
 
 #if !os(Android)
