+++ Tests/VeinTests/Predicate/ModelLevelPredicate.swift
@@ -16,9 +16,11 @@ import Logging
 import SQLiteDB
 @testable import Vein
 #if TEST_SWIFTUI
-    @testable import VeinSwiftUI
-#elseif !TEST_SWIFTUI
-    @testable import VeinCore
+@_spi(VeinTesting) @testable import VeinSwiftUI
+#elseif TEST_SCUI
+@_spi(VeinTesting) @testable import VeinSCUI
+#else
+@_spi(VeinTesting) @testable import VeinCore
 #endif
 
 @Suite
