+++ Tests/VeinTests/Relationships/RelationshipMigrationTest.swift
@@ -15,11 +15,12 @@ import Testing
 import Logging
 @testable import Vein
 #if TEST_SWIFTUI
-    @testable import VeinSwiftUI
-#elseif !TEST_SWIFTUI
-    @testable import VeinCore
+@_spi(VeinTesting) @testable import VeinSwiftUI
+#elseif TEST_SCUI
+@_spi(VeinTesting) @testable import VeinSCUI
+#else
+@_spi(VeinTesting) @testable import VeinCore
 #endif
-
 @MainActor
 @Suite struct RelationshipTest {
     static let logger = Logger(label: "de.amethystsoft.vein.test.relationship")
