+++ Tests/VeinTests/Relationships/RelationshiptDeleteRuleTest.swift
@@ -15,9 +15,11 @@ import Testing
 import Logging
 @testable import Vein
 #if TEST_SWIFTUI
-    @_spi(VeinTesting) @testable import VeinSwiftUI
-#elseif !TEST_SWIFTUI
-    @_spi(VeinTesting) @testable import VeinCore
+@_spi(VeinTesting) @testable import VeinSwiftUI
+#elseif TEST_SCUI
+@_spi(VeinTesting) @testable import VeinSCUI
+#else
+@_spi(VeinTesting) @testable import VeinCore
 #endif
 
 @MainActor
